z_reference/Full_Project_Backup/requirements.txt
python-dotenv
pandas
psycopg2-binary
pyyaml
reportlab
openpyxl
requests
rich
holidays
pymupdf
pypdf
pillow
